+++ docs/changelog.rst
@@ -5,7 +5,11 @@ Changelog
 Changes in dev
 ==============
 
+<<<<<<< HEAD
 - Added spec file for rpm-based distributions
+=======
+- Updated ComplexFields so if required they won't accept empty lists / dicts
+>>>>>>> listfield
 - Fixed ListField so it doesnt accept strings
 - Added DynamicDocument and EmbeddedDynamicDocument classes for expando schemas