+++ docs/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 Changes in dev
 ==============
 
+- Updated ComplexFields so if required they won't accept empty lists / dicts
+- Added spec file for rpm-based distributions
 - Fixed ListField so it doesnt accept strings
 - Added DynamicDocument and EmbeddedDynamicDocument classes for expando schemas